+++ service annonces/i-Foot/src/main/java/ma/emsi/iFoot/repository/AnnonceRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ma.emsi.iFoot.model.Annonce;
 
 @Repository
-public interface AnnonceRepository extends MongoRepository<Annonce, Long> {
+public interface AnnonceRepository extends MongoRepository<Annonce, Integer> {
 
 }
