+++ service annonces/i-Foot/src/main/java/ma/emsi/iFoot/repository/TerrainRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ma.emsi.iFoot.model.Terrain;
 
 @Repository
-public interface TerrainRepository extends MongoRepository<Terrain, Long> {
+public interface TerrainRepository extends MongoRepository<Terrain,Integer> {
 
 }
