+++ service annonces/i-Foot/src/main/java/ma/emsi/iFoot/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import ma.emsi.iFoot.model.User;
 
 @Repository
-public interface UserRepository extends MongoRepository<User, Long> {
+public interface UserRepository extends MongoRepository<User, Integer> {
 
 	User findByUsername(String username);
 }
